+++ lib/ductwork/orphaned_claim.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Ductwork
+  class OrphanedClaim < Ductwork::Crash; end
+end