+++ lib/ductwork/thread_crash.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Ductwork
+  class ThreadCrash < Ductwork::Crash; end
+end